Clean up unused component and update README.md.
