--- components/Headerv2.jsx
@@ -1,48 +1,0 @@
-import {
-  AppBar,
-  Box,
-  Container,
-  ButtonBase,
-  Toolbar,
-  Typography,
-} from '@mui/material';
-import Image from 'next/image';
-import Link from 'next/link';
-
-import logoImage from '../public/openseattle-500x392.png';
-
-export default function Header() {
-  return (
-    <header>
-      <AppBar color='transparent' position='sticky'>
-        <Container maxWidth='lg'>
-          <Toolbar disableGutters>
-            <ButtonBase
-              component={Link}
-              href='/'
-              sx={{
-                display: 'flex',
-                alignItems: 'center',
-                gap: 1,
-                py: 1,
-                borderRadius: 2,
-              }}
-            >
-              <Box>
-                <Typography
-                  fontFamily='oswald'
-                  fontSize={28}
-                  fontWeight={400}
-                  sx={{ lineHeight: 1 }}
-                  component='h1'
-                >
-                  Open Seattle
-                </Typography>
-              </Box>
-            </ButtonBase>
-          </Toolbar>
-        </Container>
-      </AppBar>
-    </header>
-  );
-}
